worker/requirements.txt
grpcio
grpcio-tools
protobuf
statsd
capytaine
